Rename casts event type to cast
